fix(tui): ensure consistent prefix/indent between graduation and display
Root cause: render_complete_blocks called render_block_to_lines for each
block, resetting first_content_line for each. This caused every block to
get ' ● ' prefix, while render_assistant_blocks shared one flag.

Fix: Use render_assistant_blocks directly in render_complete_blocks to
ensure identical rendering between graduated content and displayed content.
Removed render_block_to_lines as it's no longer needed.

Also includes minor formatting cleanup and type fixes (content_width as usize).

+++ crates/crucible-cli/src/tui/testing/harness.rs
@@ -726,7 +726,7 @@ impl StreamingHarness {
         use crate::tui::constants::UiConstants;
         use crate::tui::graduation::check_graduation;
 
-        let content_width = UiConstants::content_width(self.harness.width) as usize;
+        let content_width = UiConstants::content_width(self.harness.width);
 
         // Use the shared graduation logic (same as runner)
         let (all_lines, result) = check_graduation(
@@ -751,7 +751,7 @@ impl StreamingHarness {
         use crate::tui::constants::UiConstants;
         use crate::tui::graduation::render_all_lines;
 
-        let content_width = UiConstants::content_width(self.harness.width) as usize;
+        let content_width = UiConstants::content_width(self.harness.width);
         let items: Vec<_> = self
             .harness
             .view
@@ -769,7 +769,7 @@ impl StreamingHarness {
         use crate::tui::constants::UiConstants;
         use crate::tui::graduation::render_all_lines;
 
-        let content_width = UiConstants::content_width(self.harness.width) as usize;
+        let content_width = UiConstants::content_width(self.harness.width);
         let items: Vec<_> = self
             .harness
             .view
